lalafo_bot/requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
asyncpg==0.30.0
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
frozenlist==1.7.0
greenlet==3.2.4
idna==3.10
kombu==5.5.4
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.4
packaging==25.0
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.43
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.1
asgiref==3.7.2
